Refactor solution for day 7
Inspired by other solutions, I use the fact that the tree is traversed
in a depth first manner. With this information I could add all sizes
from the files to the last value in the stack. Moreover, I pop the
values, add the tuple into a "final_folders" `Vec`, and lastly add the
size of the popped value to the parent every time we do a `$ cd ..`. The
end result is a `Vec<(size, name)>` for all the directories in the tree.