Refactor day 15
Found a much smarter way to implement this by looking around in Reddit.

For part one we just check on the given row where we are looking:
* If a point is within any of the sensors beacon distance, add one to
  the counter.
* If a beacon is at this given point, subtract one from the counter.

For part two we check every point of every sensors border (if the point
is within the given border by the assignment). For every of these points
we check:
* If the point is within the beacon distance of any sensor, then we know
  that it is not the point we are looking for. Continue to the next
  point.
* Otherwise we calculate the tuning frequency and break the program
  (should only exist one beacon satisfying the requirements).