Add some cool bitwise operations to day 24
I saw this on Reddit from @SLiV9 and wanted to implement it and try to
understand it. We can represent every row as an u128 since the input is
less than 128 characters wrong. We have four lists for the blizzards,
one list for the walls and one list for the presence of our player.
Every entry represents the bit state of that row (e.g. 1 represents that
the specific blizzard or wall is present at that position). For every
step we move the player all four directions and also keep it at the
current position. If the new position is invalid (wall or blizzard), we
remove it. The solution uses a lot of bitwise operations magic that took
a while to understand, but is very beautiful.