Add solution for day 19
The second part was though. It seems like the shortest solution might be
the only solution. If I try all possible paths it is too slow to run. I
backtracked from the end until we go to 'e' and as soon we get that, I
return the steps it took to get there. This happened to be the right
solution. Before I used a HashMap, and that always gave me the same
answer (or was taking too long before I cancelled it) even with the
order randomness in a HashMap suggesting there might only be one
solution.