Update part 2 for day 5 2023
Instead of brute forcing for all the seeds. We can do a smarter brute
force by going backwards. For every number we check if the destination
mapped backwards to the original source is within the ranges from the
seed. If it is return that destination, otherwise we try for destination
+ 1. This is not optimal but makes the solution run under one second
versus 200 seconds with the old solution.