Update day 5 solution to be a lot more efficient
Instead of using all the seeds inside the ranges, we just check the
start and end of ranges since we have a linear mapping. If there are
ranges in the seed ranges that are outside the source ranges we split
them up into new sub ranges and process them separately.