Update solution for aoc23 d14 with bitwise magic
We can represent the grid with two vectors of u128. One for the round
rocks and one for the squared rocks. We can do this after verifying that
the puzzle input have less than 128 columns. With this we can do some
bitwise magic to shift the rocks in different directions.