Add some optimizations for first two days of 2024
For day 1 we simply use the `abs_diff` instead of casting to i32. For
day 2 we change the `is_safe` function to a `is_safe_iter` function.
This prevents us from cloning a lot in part 2. Moreover, we change the
logic to check for all things in a loop instead of using three
`windows(2)` iterators for our checks.