Optimize solution for aoc 24 day 6
The main optimization is to do the Floyd's tortoise and hare algorithm
to find cycles. Moreover, we are getting rid of the `HashSet` and
utilizing a `Vec<Vec<bool>>` instead. Lastly, we only try to put
obstacle on the original path from part 1 as putting obstacles anywhere
else would not change anything.