Optimize solution for aoc24 day 22
For the step function with the secret, we are using some bitwise
operators since all the instructions where a power of 2. Moreover, we
collect all the total costs directly in the first loop over the numbers
instead of doing the same loop again later on for all the different
sequences. This speeds up the program quite a bit.