Add solution for aoc 2024 day 24
This was basically just reverse engineering for part 2. These are the
helper functions used to find which wires needed to be swapped. Then it
was swapped manually and once verified for a bunch of different `x` and
`y` values we sort the wire names we swapped and returns the output.
Will try and work on a general solution for this.